profile_user right management

+++ front/profile_user.form.php
@@ -36,35 +36,22 @@ $right = new Profile_User();
 $user = new User();
 
 if (isset($_POST["add"])) {
-   if (!isset($_POST['profiles_id']) || $_POST['profiles_id'] <= 0
-       || !isset($_POST['entities_id']) || $_POST['entities_id'] < 0
-       || !isset($_POST['users_id']) || $_POST['users_id'] < 0) {
 
-      addMessageAfterRedirect($LANG['common'][24],false,ERROR);
-      glpi_header($_SERVER['HTTP_REFERER']);
+   $right->check(-1,'w',$_POST);
+   if ($right->add($_POST)) {
+      Event::log($_POST["users_id"], "users", 4, "setup",
+                  $_SESSION["glpiname"]." ".$LANG['log'][61]);
    }
-   // TODO right check on User not on Profile_User
-   if ($user->can($_POST['users_id'],'r')
-       && Profile::currentUserHaveMoreRightThan(array($_POST['profiles_id'] => $_POST['profiles_id']))
-       && haveAccessToEntity($_POST['entities_id'])) {
-
-      if ($right->add($_POST)) {
-         Event::log($_POST["users_id"], "users", 4, "setup",
-                    $_SESSION["glpiname"]." ".$LANG['log'][61]);
-      }
-      glpi_header($_SERVER['HTTP_REFERER']);
-   } else {
-      displayRightError();
-   }
+   glpi_header($_SERVER['HTTP_REFERER']);
 
 } else if (isset($_POST["delete"])) {
-   // TODO right check on User not on Profile_User
-   checkRight("user","w");
 
    if (isset($_POST["item"]) && count($_POST["item"])) {
       foreach ($_POST["item"] as $key => $val) {
          if ($val == 1) {
-            $right->delete(array('id' => $key));
+            if ($right->can($key,'w')) {
+               $right->delete(array('id' => $key));
+            }
          }
       }
       if (isset($_POST["entities_id"])) {
@@ -79,12 +66,13 @@ if (isset($_POST["add"])) {
    glpi_header($_SERVER['HTTP_REFERER']);
 
 } else if (isset($_POST["moveentity"])) {
-   checkRight("user","w");
    if (isset($_POST['entities_id']) && $_POST['entities_id'] >= 0) {
       foreach ($_POST["item"] as $key => $val) {
          if ($val == 1) {
-            $right->update(array('id'          => $key,
-                                 'entities_id' => $_POST['entities_id']));
+            if ($right->can($key,'w')) {
+               $right->update(array('id'          => $key,
+                                    'entities_id' => $_POST['entities_id']));
+            }
          }
       }
    }
